backend/requirements.txt
# SEO Agent Backend Dependencies

# LangGraph and LangChain
langgraph>=0.2.6
langchain>=0.3.19
langchain-google-genai
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph-sdk>=0.1.57

# FastAPI and Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.20
httpx>=0.27.0

# Database
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg2-binary>=2.9.9
aiosqlite>=0.21.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
cryptography>=44.0.3
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0
dnspython>=2.0.0

# Task Queue and Caching
redis>=5.0.3
celery[redis]>=5.3.6

# AI Services
google-generativeai>=0.8.0
openai>=1.51.0
anthropic>=0.39.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.3
aiohttp>=3.12.9
pandas>=2.2.1
numpy>=1.26.4
python-dotenv>=1.1.0
pytrends>=4.9.2

# Content Management & Text Analysis
scikit-learn>=1.3.0
mecab-python3>=1.0.6
Janome>=0.5.0

# Development
pytest>=8.4.0
pytest-asyncio>=1.0.0
pytest-cov>=5.0.0
black>=24.2.0
ruff>=0.6.1
mypy>=1.11.1

# LangChain and LangGraph for production
langchain-core>=0.3.0
langgraph>=0.4.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0